django/requirements.txt
amqp===2.3.2  
billiard===3.5.0.4
celery===4.2.1  
Django===2.1.3
django-allauth===0.42.0
django-celery-results===1.0.4  
django-cors-headers===2.4.0
django-guardian===1.4.9  
django-mysql===3.3.0
django-rest-swagger===2.2.0
django-simple-history===2.5.1  
django-wkhtmltopdf===3.2.0  
djangorestframework===3.9.0  
djangorestframework-jwt===1.11.0 
Jinja2===2.10   
jsonfield===2.0.2  
kombu===4.2.1  
MarkupSafe===1.1.0  
mysqlclient===1.4.6
pip===18.1   
PyJWT===1.6.4  
python-dateutil===2.7.5  
python-magic===0.4.15 
pytz===2018.7 
PyYAML===3.13   
setuptools===40.6.2 
six===1.11.0 
SQLAlchemy===1.2.14
razorpay===0.2.0
vine===1.1.4  
wheel===0.32.2 
xlrd===1.1.0
PyPDF2===1.26.0
pubnub==4.1.2
